Move components into separate folder
