+++ src/explore-education-statistics-frontend/src/components/application-insights-context-provider.tsx
