+++ src/explore-education-statistics-frontend/src/components/navigation-events.ts
